Revert "Remove minimum of 10 buffers in constructor, some cameras only allow 2"
This reverts commit 6e9ff686c69a8c616fd2ed81d56f6e09a37d5590.

Only being allowed 2 buffers was not due to the camera, but due to too little RAM.
The system only had 4 GB, and Spinnaker only allowed 2 buffers.
But the camera did not actually work correctly with only 2 buffers.
The solution is to add more RAM, and then 10 buffer minimum is OK.